+++ packages/grid_client/scripts/orchestrators/caprover_worker.ts
@@ -1,19 +1,18 @@
-import { FilterOptions, MachinesModel } from "../src";
-import { config, getClient } from "./client_loader";
-import { log } from "./utils";
+import { FilterOptions, MachinesModel } from "../../src";
+import { config, getClient } from "../client_loader";
+import { log } from "../utils";
 
 async function deploy(client, vms) {
   const res = await client.machines.deploy(vms);
-  log("================= Deploying Caprover cluster =================");
+  log("================= Deploying Worker =================");
   log(res);
-  log("================= Deploying Caprover cluster =================");
+  log("================= Deploying Worker =================");
 }
 
 async function getDeployment(client, vms) {
   const res = await client.machines.getObj(vms);
   log("================= Getting deployment information =================");
   log(res);
-  log(`You can access Caprover via the browser using: http://captain.${res[0].env.CAPROVER_ROOT_DOMAIN}`);
   log("================= Getting deployment information =================");
 }
 
@@ -43,12 +42,12 @@ async function main() {
     },
     machines: [
       {
-        name: "testvm",
-        node_id: +(await grid3.capacity.filterNodes(vmQueryOptions))[0].nodeId,
+        name: "capworker1",
+        node_id: +(await grid3.capacity.filterNodes(vmQueryOptions))[1].nodeId,
         disks: [
           {
             name: "wedDisk",
-            size: 8,
+            size: 10,
             mountpoint: "/var/lib/docker",
           },
         ],
@@ -62,19 +61,19 @@ async function main() {
         flist: "https://hub.grid.tf/tf-official-apps/tf-caprover-latest.flist",
         entrypoint: "/sbin/zinit init",
         env: {
+          // These env. vars needed to be changed based on the leader node.
           PUBLIC_KEY: config.ssh_key,
-          SWM_NODE_MODE: "leader",
-          CAPROVER_ROOT_DOMAIN: "rafy.grid.tf", // update me
-          DEFAULT_PASSWORD: "captain42",
+          SWM_NODE_MODE: "worker",
+          LEADER_PUBLIC_IP: "185.206.122.157",
           CAPTAIN_IMAGE_VERSION: "latest",
         },
       },
     ],
     metadata: "",
-    description: "caprover leader machine/node",
+    description: "caprover worker machine/node",
   };
 
-  //Deploy Caprover cluster
+  //Deploy Caprover worker
   await deploy(grid3, vms);
 
   //Get the deployment
